+++ src/source_distribution.rs
@@ -2,15 +2,21 @@ use crate::module_writer::{add_data, ModuleWriter};
 use crate::pyproject_toml::SdistGenerator;
 use crate::{pyproject_toml::Format, BuildContext, PyProjectToml, SDistWriter};
 use anyhow::{bail, Context, Result};
+use cargo_metadata::camino::Utf8Path;
 use cargo_metadata::{Metadata, MetadataCommand, PackageId};
 use fs_err as fs;
 use ignore::overrides::Override;
 use normpath::PathExt as _;
 use path_slash::PathExt as _;
 use std::collections::HashMap;
+use std::ffi::OsStr;
+use std::io::Write;
+use std::io::copy;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::str;
+use tempfile::NamedTempFile;
+use toml_edit::DocumentMut;
 use tracing::debug;
 
 /// Path dependency information.
@@ -45,13 +51,14 @@ fn parse_toml_file(path: &Path, kind: &str) -> Result<toml_edit::DocumentMut> {
 /// We only want to add path dependencies that are actually used
 /// to reduce the size of the source distribution.
 fn rewrite_cargo_toml(
-    manifest_path: impl AsRef<Path>,
+    document: &mut DocumentMut,
+    manifest_path: &Path,
     known_path_deps: &HashMap<String, PathDependency>,
-) -> Result<String> {
-    let manifest_path = manifest_path.as_ref();
-    debug!("Rewriting Cargo.toml at {}", manifest_path.display());
-    let mut document = parse_toml_file(manifest_path, "Cargo.toml")?;
-
+) -> Result<()> {
+    debug!(
+        "Rewriting Cargo.toml `workspace.members` at {}",
+        manifest_path.display()
+    );
     // Update workspace members
     if let Some(workspace) = document.get_mut("workspace").and_then(|x| x.as_table_mut()) {
         if let Some(members) = workspace.get_mut("members").and_then(|x| x.as_array()) {
@@ -100,7 +107,34 @@ fn rewrite_cargo_toml(
             }
         }
     }
-    Ok(document.to_string())
+    Ok(())
+}
+
+/// Rewrite `Cargo.toml` to find the readme in the same directory.
+///
+/// `package.readme` may point to any point above the package, so when we move the directory, but
+/// keep the readme position, we could get different readme files at the same archive location.
+/// Putting the readme in the same directory as the `Cargo.toml` prevents this.
+fn rewrite_cargo_toml_readme(
+    document: &mut DocumentMut,
+    manifest_path: &Path,
+    readme_name: Option<&str>,
+) -> Result<()> {
+    debug!(
+        "Rewriting Cargo.toml `package.readme` at {}",
+        manifest_path.display()
+    );
+
+    if let Some(readme_name) = readme_name {
+        let project = document.get_mut("package").with_context(|| {
+            format!(
+                "Missing `[package]` table in Cargo.toml with readme at {}",
+                manifest_path.display()
+            )
+        })?;
+        project["readme"] = toml_edit::value(readme_name);
+    }
+    Ok(())
 }
 
 /// When `pyproject.toml` is inside the Cargo workspace root,
@@ -148,13 +182,15 @@ fn add_crate_to_source_distribution(
     writer: &mut SDistWriter,
     manifest_path: impl AsRef<Path>,
     prefix: impl AsRef<Path>,
+    readme: Option<&Path>,
     known_path_deps: &HashMap<String, PathDependency>,
     root_crate: bool,
     skip_cargo_toml: bool,
 ) -> Result<()> {
     let manifest_path = manifest_path.as_ref();
+    let args = ["package", "--list", "--allow-dirty", "--manifest-path"];
     let output = Command::new("cargo")
-        .args(["package", "--list", "--allow-dirty", "--manifest-path"])
+        .args(args)
         .arg(manifest_path)
         .output()
         .with_context(|| {
@@ -171,6 +207,10 @@ fn add_crate_to_source_distribution(
             String::from_utf8_lossy(&output.stdout),
             String::from_utf8_lossy(&output.stderr),
         );
+    }
+    if !output.stderr.is_empty() {
+        eprintln!("From `cargo {}`:", args.join(" "));
+        std::io::stderr().write_all(&output.stderr)?;
     }
 
     let file_list: Vec<&Path> = str::from_utf8(&output.stdout)
@@ -216,15 +256,33 @@ fn add_crate_to_source_distribution(
 
     let cargo_toml_path = prefix.join(manifest_path.file_name().unwrap());
 
+    let readme_name = readme
+        .as_ref()
+        .map(|readme| {
+            readme
+                .file_name()
+                .and_then(OsStr::to_str)
+                .with_context(|| format!("Missing readme filename for {}", manifest_path.display()))
+        })
+        .transpose()?;
+
     if root_crate {
-        let rewritten_cargo_toml = rewrite_cargo_toml(manifest_path, known_path_deps)?;
+        let mut document = parse_toml_file(manifest_path, "Cargo.toml")?;
+        rewrite_cargo_toml_readme(&mut document, manifest_path, readme_name)?;
+        rewrite_cargo_toml(&mut document, manifest_path, known_path_deps)?;
         writer.add_bytes(
             cargo_toml_path,
             Some(manifest_path),
-            rewritten_cargo_toml.as_bytes(),
+            document.to_string().as_bytes(),
         )?;
     } else if !skip_cargo_toml {
-        writer.add_file(cargo_toml_path, manifest_path)?;
+        let mut document = parse_toml_file(manifest_path, "Cargo.toml")?;
+        rewrite_cargo_toml_readme(&mut document, manifest_path, readme_name)?;
+        writer.add_bytes(
+            cargo_toml_path,
+            Some(manifest_path),
+            document.to_string().as_bytes(),
+        )?;
     }
 
     for (target, source) in target_source {
@@ -399,52 +457,17 @@ fn add_cargo_package_files_to_sdist(
 
     // Add local path dependencies
     for (name, path_dep) in known_path_deps.iter() {
-        debug!(
-            "Adding path dependency: {} at {}",
-            name,
-            path_dep.manifest_path.display()
-        );
-        let path_dep_manifest_dir = path_dep.manifest_path.parent().unwrap();
-        let relative_path_dep_manifest_dir =
-            path_dep_manifest_dir.strip_prefix(&sdist_root).unwrap();
-        // we may need to rewrite workspace Cargo.toml later so don't add it to sdist yet
-        let skip_cargo_toml = workspace_manifest_path == path_dep.manifest_path;
-        add_crate_to_source_distribution(
+        add_path_dep(
             writer,
-            &path_dep.manifest_path,
-            root_dir.join(relative_path_dep_manifest_dir),
+            root_dir,
+            workspace_root,
+            &workspace_manifest_path,
             &known_path_deps,
-            false,
-            skip_cargo_toml,
+            &sdist_root,
+            name,
+            path_dep,
         )
-        .with_context(|| {
-            format!(
-                "Failed to add local dependency {} at {} to the source distribution",
-                name,
-                path_dep.manifest_path.display()
-            )
-        })?;
-        // Handle possible relative readme field in Cargo.toml
-        if let Some(readme) = path_dep.readme.as_ref() {
-            let readme = path_dep_manifest_dir.join(readme);
-            let abs_readme = readme
-                .normalize()
-                .with_context(|| format!("failed to normalize readme path `{}`", readme.display()))?
-                .into_path_buf();
-            let relative_readme = abs_readme.strip_prefix(&sdist_root).unwrap();
-            writer.add_file(root_dir.join(relative_readme), &abs_readme)?;
-        }
-        // Handle different workspace manifest
-        if &path_dep.workspace_root != workspace_root {
-            let path_dep_workspace_manifest = path_dep.workspace_root.join("Cargo.toml");
-            let relative_path_dep_workspace_manifest = path_dep_workspace_manifest
-                .strip_prefix(&sdist_root)
-                .unwrap();
-            writer.add_file(
-                root_dir.join(relative_path_dep_workspace_manifest),
-                &path_dep_workspace_manifest,
-            )?;
-        }
+        .with_context(|| format!("Failed to add path dependency {}", name))?;
     }
 
     // Add the main crate
@@ -468,6 +491,7 @@ fn add_cargo_package_files_to_sdist(
         writer,
         manifest_path,
         root_dir.join(relative_main_crate_manifest_dir),
+        None,
         &known_path_deps,
         true,
         false,
@@ -479,8 +503,9 @@ fn add_cargo_package_files_to_sdist(
             .normalize()
             .with_context(|| format!("failed to normalize readme path `{}`", readme.display()))?
             .into_path_buf();
-        let relative_readme = abs_readme.strip_prefix(&sdist_root).unwrap();
-        writer.add_file(root_dir.join(relative_readme), &abs_readme)?;
+        // Add readme next to Cargo.toml so we don't get collisions between crates using readmes
+        // higher up the file tree.
+        writer.add_file(root_dir.join(readme.file_name().unwrap()), &abs_readme)?;
     }
 
     // Add Cargo.lock file and workspace Cargo.toml
@@ -524,11 +549,17 @@ fn add_cargo_package_files_to_sdist(
                     readme: None,
                 },
             );
-            let workspace_cargo_toml = rewrite_cargo_toml(&workspace_manifest_path, &deps_to_keep)?;
+            let mut document =
+                parse_toml_file(workspace_manifest_path.as_std_path(), "Cargo.toml")?;
+            rewrite_cargo_toml(
+                &mut document,
+                workspace_manifest_path.as_std_path(),
+                &deps_to_keep,
+            )?;
             writer.add_bytes(
                 root_dir.join(relative_workspace_cargo_toml),
                 Some(workspace_manifest_path.as_std_path()),
-                workspace_cargo_toml.as_bytes(),
+                document.to_string().as_bytes(),
             )?;
         }
     } else if cargo_lock_required {
@@ -594,6 +625,72 @@ fn add_cargo_package_files_to_sdist(
         }
     }
 
+    Ok(())
+}
+
+#[allow(clippy::too_many_arguments)] // TODO(konsti)
+fn add_path_dep(
+    writer: &mut SDistWriter,
+    root_dir: &Path,
+    workspace_root: &Utf8Path,
+    workspace_manifest_path: &Utf8Path,
+    known_path_deps: &HashMap<String, PathDependency>,
+    sdist_root: &Path,
+    name: &str,
+    path_dep: &PathDependency,
+) -> Result<()> {
+    debug!(
+        "Adding path dependency: {} at {}",
+        name,
+        path_dep.manifest_path.display()
+    );
+    let path_dep_manifest_dir = path_dep.manifest_path.parent().unwrap();
+    let relative_path_dep_manifest_dir = path_dep_manifest_dir.strip_prefix(sdist_root).unwrap();
+    // we may need to rewrite workspace Cargo.toml later so don't add it to sdist yet
+    let skip_cargo_toml = workspace_manifest_path == path_dep.manifest_path;
+    add_crate_to_source_distribution(
+        writer,
+        &path_dep.manifest_path,
+        root_dir.join(relative_path_dep_manifest_dir),
+        path_dep.readme.as_deref(),
+        known_path_deps,
+        false,
+        skip_cargo_toml,
+    )
+    .with_context(|| {
+        format!(
+            "Failed to add local dependency {} at {} to the source distribution",
+            name,
+            path_dep.manifest_path.display()
+        )
+    })?;
+    // Include readme
+    if let Some(readme) = path_dep.readme.as_ref() {
+        let readme = path_dep_manifest_dir.join(readme);
+        let abs_readme = readme
+            .normalize()
+            .with_context(|| format!("failed to normalize readme path `{}`", readme.display()))?
+            .into_path_buf();
+        // Add readme next to Cargo.toml so we don't get collisions between crates using readmes
+        // higher up the file tree. See also [`rewrite_cargo_toml_readme`].
+        writer.add_file(
+            root_dir
+                .join(relative_path_dep_manifest_dir)
+                .join(readme.file_name().unwrap()),
+            &abs_readme,
+        )?;
+    }
+    // Handle different workspace manifest
+    if path_dep.workspace_root != workspace_root {
+        let path_dep_workspace_manifest = path_dep.workspace_root.join("Cargo.toml");
+        let relative_path_dep_workspace_manifest = path_dep_workspace_manifest
+            .strip_prefix(sdist_root)
+            .unwrap();
+        writer.add_file(
+            root_dir.join(relative_path_dep_workspace_manifest),
+            &path_dep_workspace_manifest,
+        )?;
+    }
     Ok(())
 }