+++ src/target.rs
@@ -1,4 +1,6 @@
 use crate::cross_compile::is_cross_compiling;
+use crate::python_interpreter::InterpreterKind;
+use crate::python_interpreter::InterpreterKind::{CPython, GraalPy, PyPy};
 use crate::PlatformTag;
 use anyhow::{anyhow, bail, format_err, Result};
 use platform_info::*;
@@ -365,6 +367,44 @@ impl Target {
             Arch::Mipsel | Arch::Mips => "mips",
             Arch::Sparc64 => "sparc64",
             Arch::LoongArch64 => "loongarch64",
+        }
+    }
+
+    /// Returns the extension architecture name python uses in `ext_suffix` for this architecture.
+    pub fn get_python_ext_arch(&self, python_impl: InterpreterKind) -> &str {
+        if matches!(self.target_arch(), Arch::Armv6L | Arch::Armv7L) {
+            "arm"
+        } else if matches!(self.target_arch(), Arch::Powerpc64Le)
+            && python_impl == InterpreterKind::PyPy
+        {
+            "ppc_64"
+        } else if matches!(self.target_arch(), Arch::X86) && python_impl == InterpreterKind::PyPy {
+            "x86"
+        } else if matches!(self.target_arch(), Arch::Powerpc) {
+            "powerpc"
+        } else {
+            self.get_python_arch()
+        }
+    }
+
+    /// Returns the environment python uses in `ext_suffix` for this architecture.
+    pub fn get_python_target_env(
+        &self,
+        python_impl: InterpreterKind,
+        python_version: (usize, usize),
+    ) -> String {
+        match python_impl {
+            CPython => {
+                // For musl handling see https://github.com/pypa/auditwheel/issues/349
+                if matches!(self.target_arch(), Arch::Mips64 | Arch::Mips64el) && self.is_linux() {
+                    "gnuabi64".to_string()
+                } else if python_version >= (3, 11) {
+                    self.target_env().to_string()
+                } else {
+                    self.target_env().to_string().replace("musl", "gnu")
+                }
+            }
+            PyPy | GraalPy => "gnu".to_string(),
         }
     }