+++ src/upload.rs
@@ -59,8 +59,9 @@ pub struct PublishOpt {
 }
 
 impl PublishOpt {
-    const DEFAULT_REPOSITORY_URL: &'static str = "https://upload.pypi.org/legacy/";
-    const TEST_REPOSITORY_URL: &'static str = "https://test.pypi.org/legacy/";
+    // Here we omit trailing slashes from the repository URL, which we'll add back in `complete_registry`
+    const DEFAULT_REPOSITORY_URL: &'static str = "https://upload.pypi.org/legacy";
+    const TEST_REPOSITORY_URL: &'static str = "https://test.pypi.org/legacy";
 
     /// Set to non interactive mode if we're running on CI
     pub fn non_interactive_on_ci(&mut self) {
@@ -316,20 +317,33 @@ fn complete_registry(opt: &PublishOpt) -> Result<Registry> {
     let pypirc = load_pypirc();
     let (registry_name, registry_url) = if let Some(repository_url) = opt.repository_url.as_deref()
     {
-        let name = match repository_url {
-            PublishOpt::DEFAULT_REPOSITORY_URL => Some("pypi"),
-            PublishOpt::TEST_REPOSITORY_URL => Some("testpypi"),
-            _ => None,
-        };
-        (name, repository_url.to_string())
+        // to normalize URLs by removing trailing slashes
+        match repository_url.trim_end_matches('/') {
+            PublishOpt::DEFAULT_REPOSITORY_URL => (
+                Some("pypi"),
+                // Add trailing slash back
+                format!("{}/", PublishOpt::DEFAULT_REPOSITORY_URL),
+            ),
+            PublishOpt::TEST_REPOSITORY_URL => (
+                Some("testpypi"),
+                // Add trailing slash back
+                format!("{}/", PublishOpt::TEST_REPOSITORY_URL),
+            ),
+            _ => (None, repository_url.to_string()),
+        }
     } else if let Some(url) = pypirc.get(&opt.repository, "repository") {
         (Some(opt.repository.as_str()), url)
     } else if opt.repository == "pypi" {
-        (Some("pypi"), PublishOpt::DEFAULT_REPOSITORY_URL.to_string())
+        (
+            Some("pypi"),
+            // Add trailing slash back
+            format!("{}/", PublishOpt::DEFAULT_REPOSITORY_URL),
+        )
     } else if opt.repository == "testpypi" {
         (
             Some("testpypi"),
-            PublishOpt::TEST_REPOSITORY_URL.to_string(),
+            // Add trailing slash back
+            format!("{}/", PublishOpt::TEST_REPOSITORY_URL),
         )
     } else {
         bail!(