+++ test-crates/uniffi-multiple-binding-files/src/lib.rs
@@ -1,0 +1,6 @@
+#[uniffi::export]
+pub fn get_status() -> mylib::Status {
+    mylib::get_status()
+}
+
+uniffi::setup_scaffolding!();