+++ Changelog.md
@@ -1,6 +1,25 @@
 # Changelog
 
-## [Unreleased]
+## [1.7.4]
+
+* Fix musllinux rpath for non-cffi bindings in [#2233](https://github.com/PyO3/maturin/pull/2233)
+* Add GitHub Actions attestation support to generate-ci in [#2234](https://github.com/PyO3/maturin/pull/2234)
+
+## [1.7.3]
+
+* Fix upload regression to pypi/testpypi in [#2229](https://github.com/PyO3/maturin/pull/2229)
+
+## [1.7.2]
+
+* Fix cross compilation issues for armv7l, mips64 and ppc in [#2204](https://github.com/PyO3/maturin/pull/2204)
+* UniFFI: supports bindings generated from multiple crates in [#2208](https://github.com/PyO3/maturin/pull/2208)
+* Enable `--all-features` when building source distribution in [#2215](https://github.com/PyO3/maturin/pull/2215)
+* Fix rpath when module-name contains `.` in [#2219](https://github.com/PyO3/maturin/pull/2219)
+
+## [1.7.1]
+
+* Forward `cargo package --list` warnings in [#2186](https://github.com/PyO3/maturin/pull/2186)
+* In source distributions, we move the readmes of path dependencies into the respective crate to avoid collision between different readmes in [#2184](https://github.com/PyO3/maturin/pull/2184)
 
 ## [1.7.0] - 2024-07-07
 
@@ -973,7 +992,11 @@ points-0.1.0-py2.py3-none-manylinux1_x86_64.whl | 2,8M | 752K | 85K
 
  * Initial Release
 
-[Unreleased]: https://github.com/pyo3/maturin/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/pyo3/maturin/compare/v1.7.4...HEAD
+[1.7.4]: https://github.com/pyo3/maturin/compare/v1.7.3...v1.7.4
+[1.7.3]: https://github.com/pyo3/maturin/compare/v1.7.2...v1.7.3
+[1.7.2]: https://github.com/pyo3/maturin/compare/v1.7.1...v1.7.2
+[1.7.1]: https://github.com/pyo3/maturin/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/pyo3/maturin/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/pyo3/maturin/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/pyo3/maturin/compare/v1.5.0...v1.5.1